table no longer in main code tried components
