+++ components/ic_dash/store.ts
@@ -1,0 +1,8 @@
+import { Patient } from "./types";
+
+
+type State = {
+    patients: Patient[];
+    setPatients: (patients: Patient[]) => void;
+}
+
